+++ cmd/snitchbar/platform_darwin.go
@@ -5,14 +5,7 @@ package main
 import (
 	"fmt"
 	"os/exec"
-	"strings"
 )
-
-func copyToClipboard(text string) error {
-	cmd := exec.Command("pbcopy")
-	cmd.Stdin = strings.NewReader(text)
-	return cmd.Run()
-}
 
 func openTerminal(command string) error {
 	script := fmt.Sprintf(`tell application "Terminal"